clang-tidy fix

+++ tasks/krykov_e_word_count/mpi/src/ops_mpi.cpp
@@ -139,8 +139,8 @@ bool KrykovEWordCountMPI::RunImpl() {
              MPI_COMM_WORLD);
   MPI_Gather(&ends_with_space, 1, MPI_INT, world_rank == 0 ? all_ends.data() : nullptr, 1, MPI_INT, 0, MPI_COMM_WORLD);
 
-  unsigned long long local_count_ull = static_cast<unsigned long long>(local_count);
-  std::vector<unsigned long long> all_counts_ull(world_size);
+  uint64_t local_count_ull = static_cast<uint64_t>(local_count);
+  std::vector<uint64_t> all_counts_ull(world_size);
   MPI_Gather(&local_count_ull, 1, MPI_UNSIGNED_LONG_LONG, world_rank == 0 ? all_counts_ull.data() : nullptr, 1,
              MPI_UNSIGNED_LONG_LONG, 0, MPI_COMM_WORLD);
